==> Week 01/00-welcome.ipynb <==
# 👋 Welcome to your first INFOMPPM seminar 👋

During the seminar session, we will first discuss how (public) values connect to recommender systems. You will then need to think about the data you would need to build a recommender system and how that poses opportunities but also risks for different values. We then turn to the basics of building a recommender in Python:
- Non-personalised recommendations (ratings, seeded, confidence, support)
- Implicit ratings
- Running StreamlitThe activities aim for testing your knowledge about the readings, getting your code book running, extracting features from existing databases, and practising with core concepts.

### Dataset
The dataset you for this assignment will be the [Book-Crossing Dataset](http://www2.informatik.uni-freiburg.de/~cziegler/BX/) mined by [Cai-Nicolas Ziegler](http://dbis.informatik.uni-freiburg.de/team/ziegler/cai). 

The dataset:
> ... a 4-week crawl (August / September 2004) from the Book-Crossing community with kind permission from Ron Hornbaker, CTO of Humankind Systems. Contains 278,858 users (anonymized but with demographic information) providing 1,149,780 ratings (explicit / implicit) about 271,379 books.

While one could argue that the dataset is dated, it is a good starters dataset that will still offer you some challenges.
