==> Week 01/01-exploration.ipynb <==
# 🔬 Data exploration and preparation
In this notebook, you will look at the data and create a subset of that data. The dataset was "relatively" clean upon download, but the lecturers got rid of some pesky delimiter issues. If you want to encounter these issues yourself, you can use the original dataset found at the [Book-Crossing Dataset](http://www2.informatik.uni-freiburg.de/~cziegler/BX/).

### 1. Loading the data
Load the three datasets and explore the data.


import pandas as pd

#dowload the files
ratings = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/BX-Book-Ratings.csv', sep=';')
ratings.head()

books = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/BX-Books.csv', sep=';', encoding="latin-1",low_memory=False )
books.head()

users = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/BX-Users.csv', sep=';', encoding="latin-1")
users.head()

### 2. Cleaning the data
Check if all reviews are connected to a book. Is there a review but no book or user connected to this review? Check if all the authors are spelled correctly, are there outliers (e.g., users with insane number of reviews) etc etc. What were your steps in cleaning the data?

Check data overall: methods describe(), shape, NaN-values.

users.describe()

books.describe()

ratings.describe()

books.isna().sum()

ratings.isna().sum()

users.isna().sum()

ratings.shape

books.shape

users.shape

users.nunique()

books.nunique()

books['Year-Of-Publication'].value_counts() #there are some 'strange' years but for now we don't care about the year

ratings.nunique()

ratings['Book-Rating'].value_counts()

books.loc[books['Book-Title']=='Robots and Empire']

#drop diblicates of the same books from different publishers
books = books.drop_duplicates(subset=['Book-Title', 'Book-Author'], keep='last')
books.loc[books['Book-Title']=='Robots and Empire']

books.shape

ratings.shape

ratings = ratings[ratings['Book-Rating'] != 0] #drop reviews with 0-ratings

ratings.shape

# check of there are non existing ID-s in ratings
df_books_users_out = pd.merge(books, ratings, on = 'ISBN', how='outer', indicator=True)
df_books_users_out._merge.value_counts()

list1 = df_books_users_out.query("_merge=='right_only'")['ISBN']
list1.shape

ratings = ratings[ratings.ISBN.isin(list1) == False]
ratings

# check if there is non-existing users in ratings
df_users_ratings_out = pd.merge(users, ratings, how='outer', indicator=True)
df_users_ratings_out._merge.value_counts()

# spelling mistakes in authors
books['Book-Author'].value_counts() # there are Not Applicable (NA)

# outliers - insane numbers of reviews

ratings.groupby('ISBN').agg({'User-ID':'count'}).sort_values(by='User-ID', ascending=False)

ratings.shape

### 3. Subsetting the data
The publication accompanied with this dataset [Improving Recommendation Lists Through Topic Diversification](http://www2.informatik.uni-freiburg.de/~cziegler/BX/WWW-2005-Preprint.pdf) by Cai-Nicolas Ziegler, Sean M. McNee, Joseph A. Konstan, Georg Lausen; describes the process of subsetting (condensation steps) the dataset as follows (p5): 

> Hence, we discarded all books missing taxonomic descriptions, along with all ratings referring to them. Next, we also removed book titles with fewer than 20 overall mentions. Only community members with at least five ratings each were kept. 

Explore what these parameters mean for the overall dataset. Also, consider if you want the implicit ratings (Book-Rating == 0) in the final dataset. What would the implications be? Would you exclude it before the other parameters, or would you exclude them afterwards? 

While the publication describes the resulting dataset's dimensions, your results might differ. But that is ok for now.


ratings.head()

# Next, we also removed book titles with fewer than 20 overall mentions. Only community members with at least 5 ratings each were kept.
x = ratings['ISBN'].value_counts() >=20
idx = x[x].index
ratings = ratings[ratings['ISBN'].isin(idx)]

y = ratings['User-ID'].value_counts() >5
idy = y[y].index
ratings = ratings[ratings['User-ID'].isin(idy)]

#final shape of rating dataset
ratings.shape

### 4. Extra step
Take a closer look at `BX-Books.csv` and search for a book named _Robots and Empire_ by Isaac Asimov. What do you encounter? Is this something you would solve? 

Let us argue that this is problematic for our dataset. How would you solve this? You might want to redo step 2 if you choose to take this extra step.

#this step I did in cleaning step - there are diblicates (one book - a few publishers)
#books = books.drop_duplicates(subset=['Book-Title', 'Book-Author'], keep='last')
#books.loc[books['Book-Title']=='Robots and Empire']

### 5. Save the new dataset(s)
Save the dataset(s) in distinct named CSV-files for later usage. Move the file(s) to the data directory.


books.to_csv('books.csv')
ratings.to_csv('ratings.csv')
users.to_csv('users.csv')
