==> Week 01/02-most-reviewed.ipynb <==
# 📚 Recommendations based on most reviewed books
You are going to create your first recommendations based most reviewed books. While this is far from personalised, it is still widely applied, and it does offer you to explore the Streamlit app in the app directory.

### 1. Calculate the total reviews per book

import pandas as pd

#reading the files
ratings = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/ratings.csv')
books = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/books.csv')
users = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/users.csv')

#making top-10 by total
top_recommends = ratings.groupby('ISBN').agg({'Book-Rating':'sum'}).reset_index().sort_values(by='Book-Rating', ascending=False)
top_recommends = top_recommends.rename(columns={"Book-Rating":'count'})
top_recommends

#checking how many reveiws has top-10 by avg

# top_recommends = ratings.groupby('ISBN').agg({'Book-Rating':['sum','count']}).reset_index()
# # top_recommends = top_recommends.rename(columns={"Book-Rating":'count'})
# top_recommends.columns = ['ISBN', 'sum', 'count']
# top_recommends.sort_values(by='sum', ascending=False)

### 2. Save the recommendations
Select the top 10 recommendations based on amount of reviews, save the recommendations as `recommendations-most-reviewed.csv`, and replace the file in the `app/recommendations` directory. The app already has some recommendations which are in dire need of fixing. Make sure the file has the following columns: `ISBN;count`

#saving to csv
top_10_recommends = top_recommends.head(10)
top_10_recommends.to_csv('recommendations-most-reviewed.csv', sep=';', index=False)

### 3. Run the Streamlit app
Perhaps this is the first time you will run a Streamlit app. We provided you with boilerplate code to see your recommendations in a working interface. Over time you might decide to tweak a few buttons or add some metadata, so it is good to familiarise yourself with the [Streamlit documentation](https://docs.streamlit.io/library/api-reference). As a future data scientist creating quick proofs-of-concept is a necessity.

Go to the app and start the app `streamlit run app.py`
