==> Week 01/03-average-ratings.ipynb <==
# 📚 Recommendations based on average ratings
You are going to create your first recommenations based on average ratings.

### 1. Calculate the average ratings
Calculate the average ratings and the number of reviews (count) for the books in your new dataset(s).

import pandas as pd
import numpy as np

#reading files
ratings = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/ratings.csv')
books = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/books.csv')
users = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/users.csv')

# top-10 by average
top_recommends_avg = ratings.groupby('ISBN').agg({'Book-Rating':np.average}).sort_values(by='Book-Rating', ascending=False).reset_index()
top_recommends_avg_10 = top_recommends_avg.head(10)
top_recommends_avg

#checking how many reviews has top-10
# top_recommends_avg = ratings.groupby('ISBN').agg({'Book-Rating':[np.average, 'count']}).reset_index()
# top_recommends_avg.columns = ['ISBN', 'average', 'count']
# top_recommends_avg.sort_values(by='average', ascending=False)


### 2. Reflect on the recommendations
Look at the rating average of the top 10 and how many users reviewed these books. Why would this be not a suitable method of recommendation?

As we can see above, in case of average, top-10 could be designed in a way, when you get not popular (aka with fewer reviews) books at top.

### 3. Save the recommendations
Nevertheless, select the top 10 of the ratings and save the recommendations as `recommendations-ratings-avg.csv` and replace the file to the app directory. Make certain the file has the following columns: `ISBN;mean`

top_recommends_avg_10.to_csv('recommendations-ratings-avg.csv', sep=';', index=False)
