==> Week 01/04-ratings-weighted.ipynb <==
# 📚 Recommendations based on weighted ratings

With the limitations of calculating the average ratings you will now create recommendations based on the weightage average for individual books. Take a look at [Building a Recommendation System using weighted-average score](https://medium.com/@developeraritro/building-a-recommendation-system-using-weighted-hybrid-technique-75598b6be8ed) and implement this concept.

### 1. Calculate Weightage Average for Individual books average rating
Decide what value your mean vote acrosse the whole report (C) should be.



import pandas as pd
import numpy as np

#import datasets
ratings = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/ratings.csv')
books = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/books.csv')
users = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/users.csv')


ratings.shape

#making all necessary values: total amount review
v= ratings.groupby('ISBN').agg({'Book-Rating':'count'}).sort_values(by='Book-Rating', ascending=False).reset_index().rename(columns={'Book-Rating': 'total_count'})

#make merged df with main dataset review
merged_df = ratings.merge(v)

# making coolumn rating and merge again
R=ratings.groupby('ISBN').agg({'Book-Rating':np.average}).sort_values(by='Book-Rating', ascending=False).reset_index().rename(columns={'Book-Rating': 'rating'})
merged_df = merged_df.merge(R)
merged_df

C=merged_df['Book-Rating'].mean() # 7.88 #mean of rating
m=merged_df['Book-Rating'].quantile(0.70) # 17 # quantile

# final formula ((R*v)+ (C*m))/(v+m)
# Calculate all the components based on the above formula
merged_df['weight']=((merged_df.rating*merged_df.total_count)+ (C*m))/(merged_df.total_count+m)
merged_df

merged_df = merged_df.sort_values('weight', ascending=False).groupby(['ISBN','weight']).count()
merged_df = merged_df.reset_index()
merged_df #final dataset

#make new dataset with necessary first 2 columns
df_new = merged_df.iloc [:, [0, 1]]
recommend_by_weight =  df_new.head(10)
recommend_by_weight

recommend_by_weight

### 2. Save the recommendations
Select the top 10 of the ratings based on weight and save the recommendations as `recommendations-ratings-weight.csv` and replace the file to the app directory. Make certain the file has the following columns: `ISBN;weight`

# code goes here
recommend_by_weight.to_csv('recommendations-ratings-weight.csv', sep=';', index=False)

### 3. Compare recommendations based on average rating and weighted ratings
Do you see big differences when you look at the interface?

All the recommendations looks different. And that's logical because we used different approaches to get top-10.
