==> Week 01/05-seeded-recommendations-freq.ipynb <==

# 📚 Recommendations based on Frequently Reviewed Together (frequency)
An excellent way to give quick recommendations is based on Frequently Bought Together or, in this case, Frequently Reviewed Together. A way to create combinations is with `permutations` from `itertools`. For instance, the code below creates combinations from items occurring in a list.


### 1. Quick introduction to permutations

from itertools import permutations 

# items bought together
items = ['milk', 'bread', 'eggs']

# this code creates sets of 2 items from the itemset above
list(permutations(items, 2))

### 2. Count the combinations of books reviewed together
Create combinations with `permutations` and count the occurrences. Depending on your selection in the exploration step, this can take a while.

import pandas as pd

# reading the file
ratings = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/ratings.csv')
ratings.head()

#getting permutations

permutation_two = (ratings
                .groupby('User-ID')
                .apply(lambda x: pd.DataFrame(list(permutations(x['ISBN'], 2))))
                .reset_index())
permutation_two.columns.values[2] = "a"
permutation_two.columns.values[3] = "b"
permutation_two

#getting top-10 for each ISBN
df_new = permutation_two.groupby(['a','b']).aggregate('count').reset_index().sort_values('User-ID', ascending = False)
df_new.head()

# rename a and b to book_a and book_b, choosing the right columns
df_final_10= df_new.iloc [:, [0, 1,2]]
recommend_by_pemutation =  df_final_10.rename(columns={'a': 'book_a', 'b':'book_b', 'User-ID':'count'})
recommend_by_pemutation  = recommend_by_pemutation.groupby('book_a').head(10)
recommend_by_pemutation.shape

recommend_by_pemutation.head()

### 3. Save the recommendations
You can see that this will be a large file. Since we only need the top 10 of any given book, make sure the CSV file only contains 10 recommendations per book. Save the recommendations as `recommendations-seeded-freq.csv` and replace the file in the app directory. Do not forget to uncomment the block of code connected to this step. 


recommend_by_pemutation.to_csv('recommendations-seeded-freq.csv', sep=';', index=False)
