==> Week 01/07-collecting-metadata.ipynb <==
# ⚙️ Collecting extra metadata
In the coming weeks, you will refine this book recommender system. However, in order to do so, you need extra data. This part is not to keep you busy, though it takes some time and effort, but to make you think critically about the data you need to refine this recommender system. Luckily, for this assignment, retrieving this extra metadata is relatively easy. [Google APIs](https://www.googleapis.com/books/v1/volumes?q=isbn:0451519841) offers a way to retrieve extra information based on an ISBN. 

### 1. Quick introduction to retrieving data from an API
There are many ways to retrieve information from an API, and you probably experimented with a few already. For this case, you simply need to use the `requests` library. A quick refresher below. You can always check the [official](https://docs.python-requests.org/en/latest/user/quickstart/) documentation.

import requests
res = requests.get('https://api.github.com/events')
print(res.text)

### 2. Subset the data

Make a selection of users that have more than 100 and less than 200 reviews. This results in a manageable dataset for collection the extra metadata.

import pandas as pd
ratings = pd.read_csv('/Users/ekaterinamazur/PycharmProjects/INFOMPPM1/Week 01/data/ratings.csv')

df_users = ratings.groupby('User-ID')['ISBN'].count().reset_index(name='counts')
users = list(df_users[(df_users['counts'] > 100) & (df_users['counts'] < 200)]['User-ID'])

### 3. Retrieve the data for the books in your dataset

Make a selection of users that have more than 100 and less than 200 reviews. This results in a manageable dataset for collection the extra metadata.
Retrieve the metadata and save it in files or a database. The best is to go about this process structured. Create batches of books and per batch:
1. Retrieve the data
2. Store the data
3. Extract the data

If you are clever (and time conscious), divide the collection part with a fellow student. Also, use `time.sleep()` to not overuse Google.

# retrieve metadata
